Started the add note capabilities
Need to fix how the text appears over the sticky note (goes over the
edges/might be too long etc.
Do we need to worry about each note/website/etc. have a unique id? Or
is that not necessary for the drag and drop thing?
Also need to make the text input box prettier… and hopefully center it
as well