Added Checklist abilities
Made the font into a monospace in order to be able to limit the length
of the notes/checklist items properly.
Added checklist implementation up to 12 items (otherwise would overflow
the list area).
Dynamical check the length of the note to make sure it doesn’t get
longer than 6 lines (would otherwise overflow).
Need to: disallow the child to edit the notes and add items to list.